backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "bcrypt>=4.3.0",
    "bland>=0.3.0",
    "fastapi[standard]>=0.116.1",
    "langchain-community>=0.3.28",
    "langchain-openai>=0.3.32",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
    "websockets>=15.0.1",
]
